refactor: 로그인, 회원가입 로직을 모두 UserService로 통합
