+++ tomcat/src/main/java/nextstep/jwp/service/UserService.java
@@ -4,6 +4,7 @@ import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.LoginFailException;
 import nextstep.jwp.exception.NotFoundException;
+import nextstep.jwp.exception.RegisterFailException;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http11.session.Session;
 import org.slf4j.Logger;
@@ -17,6 +18,7 @@ public class UserService {
     private static final String EMPTY_VALUE_ERROR_MESSAGE = "정보를 모두 입력해주세요.";
     private static final String WRONG_PASSWORD_ERROR_MESSAGE = "비밀번호가 맞지 않습니다.";
     private static final String NOT_FOUND_USER_ERROR_MESSAGE = "유저를 찾을 수 없습니다.";
+    private static final String DUPLICATED_ACCOUNT_ERROR_MESSAGE = "이미 존재하는 아이디입니다.";
     private static final String EMPTY_VALUE = "";
 
     private static final Logger log = LoggerFactory.getLogger(UserService.class);
@@ -42,6 +44,30 @@ public class UserService {
         if (!user.checkPassword(password)) {
             log.warn(WRONG_PASSWORD_ERROR_MESSAGE);
             throw new LoginFailException(WRONG_PASSWORD_ERROR_MESSAGE);
+        }
+    }
+
+    public void register(Map<String, String> requestBody) {
+        String account = requestBody.getOrDefault("account", EMPTY_VALUE);
+        String email = requestBody.getOrDefault("email", EMPTY_VALUE);
+        String password = requestBody.getOrDefault("password", EMPTY_VALUE);
+        checkEmptyInput(account, email, password);
+
+        checkDuplicateAccount(account);
+        InMemoryUserRepository.save(new User(account, password, email));
+    }
+
+    private void checkEmptyInput(String account, String email, String password) {
+        if (account.isBlank() || email.isBlank() || password.isBlank()) {
+            throw new RegisterFailException(EMPTY_VALUE_ERROR_MESSAGE);
+        }
+    }
+
+    private void checkDuplicateAccount(String account) {
+        boolean isExistAccount = InMemoryUserRepository.existsByAccount(account);
+        if (isExistAccount) {
+            log.error(DUPLICATED_ACCOUNT_ERROR_MESSAGE);
+            throw new RegisterFailException(DUPLICATED_ACCOUNT_ERROR_MESSAGE);
         }
     }
